quelques changements

+++ Admin/listesDattente.aspx.vb
@@ -63,32 +63,36 @@ Partial Class Admin_listesDattente
     End Sub
 
     Protected Sub btnCreerGroupe_Click(ByVal sender As Object, ByVal e As System.EventArgs) Handles btnCreerGroupe.Click
-        Dim leGroupeAjoute As Groupe = New Groupe()
-        Dim leNoGroupeDepart As Integer = hFieldNoGroupe.Value
-        Dim leGroupeDepart As Groupe = (From unGroupe In lecontext.Groupe Where unGroupe.noGroupe = leNoGroupeDepart Select unGroupe).FirstOrDefault
-        leGroupeAjoute.Local = leGroupeDepart.Local
-        leGroupeAjoute.DateDebut = leGroupeDepart.DateDebut
-        leGroupeAjoute.DateFin = leGroupeDepart.DateFin
-        leGroupeAjoute.DateLimiteInscription = leGroupeDepart.DateLimiteInscription
-        leGroupeAjoute.AgeMinimum = leGroupeDepart.AgeMinimum
-        leGroupeAjoute.Agemaximum = leGroupeDepart.Agemaximum
-        leGroupeAjoute.Actif = False
-        leGroupeAjoute.Animateur = (From dl In lecontext.Animateur Where dl.noAnimateur = leGroupeDepart.Animateur.noAnimateur Select dl).FirstOrDefault
-        leGroupeAjoute.Cours = (From dl In lecontext.Cours Where dl.noCours = leGroupeDepart.Cours.noCours Select dl).First
-        lecontext.AddObject("Groupe", leGroupeAjoute)
-
         Dim lesInscriptions As New List(Of ListeDAttente)(From uneInscription In lecontext.ListeDAttente Where uneInscription.Accepte = True Select uneInscription)
-        For Each inscription In lesInscriptions
-            Dim lePaiement As Paiement = New Paiement()
-            lePaiement.ModePaiement = "Non payé"
-            lePaiement.noPaypal = "00000"
-            lePaiement.Membre = inscription.Membre
-            lePaiement.Groupe = leGroupeAjoute
-            lecontext.ListeDAttente.DeleteObject(inscription)
-        Next
-        lecontext.SaveChanges()
+        If lesInscriptions.Count > 0 Then
+            Dim leGroupeAjoute As Groupe = New Groupe()
+            Dim leNoGroupeDepart As Integer = hFieldNoGroupe.Value
+            Dim leGroupeDepart As Groupe = (From unGroupe In lecontext.Groupe Where unGroupe.noGroupe = leNoGroupeDepart Select unGroupe).FirstOrDefault
+            leGroupeAjoute.Local = leGroupeDepart.Local
+            leGroupeAjoute.DateDebut = leGroupeDepart.DateDebut
+            leGroupeAjoute.DateFin = leGroupeDepart.DateFin
+            leGroupeAjoute.DateLimiteInscription = leGroupeDepart.DateLimiteInscription
+            leGroupeAjoute.AgeMinimum = leGroupeDepart.AgeMinimum
+            leGroupeAjoute.Agemaximum = leGroupeDepart.Agemaximum
+            leGroupeAjoute.Actif = False
+            leGroupeAjoute.Animateur = (From dl In lecontext.Animateur Where dl.noAnimateur = leGroupeDepart.Animateur.noAnimateur Select dl).FirstOrDefault
+            leGroupeAjoute.Cours = (From dl In lecontext.Cours Where dl.noCours = leGroupeDepart.Cours.noCours Select dl).First
+            lecontext.AddObject("Groupe", leGroupeAjoute)
 
-        Server.Transfer("~/Admin/GérerLesCours.aspx")
+
+            For Each inscription In lesInscriptions
+                Dim lePaiement As Paiement = New Paiement()
+                lePaiement.ModePaiement = "Non payé"
+                lePaiement.noPaypal = "00000"
+                lePaiement.Membre = inscription.Membre
+                lePaiement.Groupe = leGroupeAjoute
+                lecontext.ListeDAttente.DeleteObject(inscription)
+            Next
+            lecontext.SaveChanges()
+
+            Server.Transfer("~/Admin/GérerLesCours.aspx")
+        End If
+        lbMessage.Text = "Aucun groupe n'a été créé puisqu'il n'y a personne à mettre dedans."
     End Sub
 
     Protected Sub btnEnregistrer_Click(ByVal sender As Object, ByVal e As System.EventArgs) Handles btnEnregistrer.Click
